+++ extensions/fetch/22_body.js
@@ -5,8 +5,7 @@
 /// <reference path="../url/internal.d.ts" />
 /// <reference path="../url/lib.deno_url.d.ts" />
 /// <reference path="../web/internal.d.ts" />
-/// <reference path="../file/internal.d.ts" />
-/// <reference path="../file/lib.deno_file.d.ts" />
+/// <reference path="../web/lib.deno_web.d.ts" />
 /// <reference path="./internal.d.ts" />
 /// <reference path="./11_streams_types.d.ts" />
 /// <reference path="./lib.deno_fetch.d.ts" />
