--- extensions/file/README.md
@@ -1,5 +1,0 @@
-# deno_file
-
-This crate implements the File API.
-
-Spec: https://w3c.github.io/FileAPI
