+++ extensions/web/README.md
@@ -1,5 +1,6 @@
 # deno web
 
-Op crate that implements Event, TextEncoder, TextDecoder.
+Op crate that implements Event, TextEncoder, TextDecoder and File API
+(https://w3c.github.io/FileAPI).
 
 Testing for text encoding is done via WPT in cli/.
