+++ extensions/web/internal.d.ts
@@ -70,5 +70,15 @@ declare namespace globalThis {
       atob(data: string): string;
       btoa(data: string): string;
     };
+
+    declare var file: {
+      Blob: typeof Blob & {
+        [globalThis.__bootstrap.file._byteSequence]: Uint8Array;
+      };
+      readonly _byteSequence: unique symbol;
+      File: typeof File & {
+        [globalThis.__bootstrap.file._byteSequence]: Uint8Array;
+      };
+    };
   }
 }
