+++ runtime/lib.rs
@@ -4,7 +4,6 @@ pub use deno_broadcast_channel;
 pub use deno_console;
 pub use deno_crypto;
 pub use deno_fetch;
-pub use deno_file;
 pub use deno_timers;
 pub use deno_url;
 pub use deno_web;
