+++ runtime/worker.rs
@@ -23,7 +23,7 @@ use deno_core::ModuleId;
 use deno_core::ModuleLoader;
 use deno_core::ModuleSpecifier;
 use deno_core::RuntimeOptions;
-use deno_file::BlobUrlStore;
+use deno_web::BlobUrlStore;
 use log::debug;
 use std::env;
 use std::pin::Pin;
@@ -96,8 +96,7 @@ impl MainWorker {
       deno_webidl::init(),
       deno_console::init(),
       deno_url::init(),
-      deno_web::init(),
-      deno_file::init(options.blob_url_store.clone(), options.location.clone()),
+      deno_web::init(options.blob_url_store.clone(), options.location.clone()),
       deno_fetch::init::<Permissions>(
         options.user_agent.clone(),
         options.ca_data.clone(),
